feat: add macro expansion
