+++ Sources/TestDRSMacros/Expectations/ConfirmationOfCallMacroExpansion.swift
@@ -1,0 +1,10 @@
+//
+// Created on 1/13/25.
+// Copyright © 2025 Turo Open Source. All rights reserved.
+//
+
+public struct ConfirmationOfCallMacro: FunctionCallVerificationMacro {
+    static let instanceVerificationName = "confirmationOfCall"
+    static let staticVerificationName = "confirmationOfCall"
+    static let functionArgumentLabel: String? = "to"
+}
